Extract EtcdHelper interface
